pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "dbutton"
version = "1.0.1"
description = "A framework-agnostic pagination and filtering library for Telegram bots"
readme = "README.md"
authors = [{name = "Ray", email = "pthndv@proton.me"}]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = []

[tool.setuptools]
packages = {find = {where = ["."]}}
# package_dir is not needed as we're using find:
# package_dir = { "" = "." }

[project.optional-dependencies]
ptb = ["python-telegram-bot>=13.0"]
aiogram = ["aiogram>=2.0"]
pyrogram = ["pyrogram>=1.0.0"]
dev = [
    "pytest>=6.0",
    "black>=21.5b2",
    "isort>=5.0",
    "mypy>=0.812",
]

[project.urls]
"Homepage" = "https://github.com/pthndv/pycypher"
"Bug Tracker" = "https://github.com/pthndv/pycypher/issues"
